Revert "BN-16 : feat/ Aggregate"
