--- src/main/java/com/example/clearthefridge/domain/ingredient/repository/IngredientRepository.java
@@ -1,14 +1,0 @@
-package com.example.clearthefridge.domain.ingredient.repository;
-
-
-import com.example.clearthefridge.domain.ingredient.entity.Ingredient;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface IngredientRepository extends JpaRepository<Ingredient, Long> {
-
-    Optional<Ingredient> findByName(String name);
-}
