--- src/main/java/com/example/clearthefridge/domain/ingredient/service/IngredientService.java
@@ -1,7 +1,0 @@
-package com.example.clearthefridge.domain.ingredient.service;
-
-
-
-public interface IngredientService {
-
-}
