--- src/main/java/com/example/clearthefridge/domain/ingredient/service/impl/IngredientServiceImpl.java
@@ -1,11 +1,0 @@
-package com.example.clearthefridge.domain.ingredient.service.impl;
-
-import com.example.clearthefridge.domain.ingredient.service.IngredientService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class IngredientServiceImpl implements IngredientService {
-
-}
