--- src/main/java/com/example/clearthefridge/domain/recipe/repository/example.java
@@ -1,4 +1,0 @@
-package com.example.clearthefridge.domain.recipe.repository;
-
-public class example {
-}
