+++ src/main/java/com/example/clearthefridge/entity/Ingredient.java
@@ -1,8 +1,7 @@
-package com.example.clearthefridge.domain.ingredient.entity;
+package com.example.clearthefridge.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -11,16 +10,11 @@ import lombok.NoArgsConstructor;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-@Builder
 public class Ingredient {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    //재료 이름
     private String name;
-
-    //소비 권장 기한
-    private Integer lifeDays;
 }
