+++ src/main/java/com/example/clearthefridge/entity/User.java
@@ -1,8 +1,7 @@
-package com.example.clearthefridge.domain.user.entity;
+package com.example.clearthefridge.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -12,7 +11,6 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
 public class User {
 
     @Id
