+++ src/main/java/com/example/clearthefridge/entity/UserIngredient.java
@@ -1,12 +1,8 @@
-package com.example.clearthefridge.domain.refrige.entity;
-import com.example.clearthefridge.domain.ingredient.entity.Ingredient;
-import com.example.clearthefridge.domain.user.entity.User;
+package com.example.clearthefridge.entity;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
 
 import java.time.LocalDateTime;
 
@@ -16,7 +12,6 @@ import java.time.LocalDateTime;
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-@Builder
 public class UserIngredient {
 
     @Id
@@ -29,18 +24,6 @@ public class UserIngredient {
     @ManyToOne
     private Ingredient ingredient;
 
-    //재료 단위 ex g, ml,
-    private String unit;
-
-    //재료 수량
     private String amount;
-
-    //유통기한 만료일
     private LocalDateTime expiryDate;
-
-    //재료 등록일
-    @CreatedDate
-    private LocalDateTime createdAt;
-
-
 }
