--- src/main/java/com/example/clearthefridge/global/config/SecurityConfig.java
@@ -1,4 +1,0 @@
-package com.example.clearthefridge.global.config;
-
-public class SecurityConfig {
-}
